+++ example/jest-setup.js
@@ -1,6 +1,0 @@
-import Enzyme from 'enzyme'
-import Adapter from 'enzyme-adapter-react-16'
-
-// other enzyme tools (not used yet)
-// - https://github.com/FormidableLabs/enzyme-matchers
-Enzyme.configure({ adapter: new Adapter() })